Refactor server.js and use Clarifai from the backend
