+++ controllers/profile.js
@@ -1,0 +1,21 @@
+const handleProfile = (req, res, db) => {
+    const { id } = req.params;
+    db.select('*').from('users').where({ id })
+        .then((user) => {
+            if (user.length) {
+                res.json(user[0])
+            }
+            else {
+                res.status(404).json('user not found');
+            }
+
+
+        }).catch((err) => {
+            res.status(500).json('Oops! Something is wrong');
+        });
+
+}
+
+module.exports = {
+    handleProfile
+};
