+++ server.js
@@ -4,6 +4,11 @@ const bcrypt = require('bcrypt-nodejs');
 const cors = require('cors');
 const knex = require('knex');
 
+const register = require('./controllers/register');
+const signin = require('./controllers/signin');
+const profile = require('./controllers/profile');
+const image = require('./controllers/image');
+
 const db = knex({
     client: 'pg',
     connection: {
@@ -19,88 +24,11 @@ const app = express();
 app.use(bodyParser.json());
 app.use(cors());
 
-app.post('/signin', (req, res) => {
-    const { email, password } = req.body;
-    db.select('email', 'hash').from('login')
-        .where('email', '=', email)
-        .then(data => {
-            const isValid = bcrypt.compareSync(password, data[0].hash);
-            if (isValid) {
-                return db.select('*').from('users')
-                    .where('email', '=', email)
-                    .then(user => {
-                        res.json(user[0])
-                    })
-                    .catch(err => res.status(500).json('unable to get the user'))
-            }
-            else {
-                res.status(400).json('wrong credentials');
-            }
-        })
-        .catch(err => res.status(400).json('Wrong credentials'))
-
-})
-
-app.post('/register', (req, res) => {
-    const { email, name, password } = req.body;
-    const hash = bcrypt.hashSync(password);
-    db.transaction(trx => {
-        trx.insert({
-            hash,
-            email
-        })
-            .into('login')
-            .returning('email')
-            .then(loginEmail => {
-                return trx('users')
-                    .returning('*')
-                    .insert({
-                        email: loginEmail[0],
-                        name,
-                        joined: new Date()
-                    })
-                    .then(user => {
-                        res.json(user[0]);
-                    })
-
-            })
-            .then(trx.commit)
-            .catch(trx.rollback)
-    })
-        .catch(err => res.status(400).json(err))
-})
-
-app.get('/profile/:id', (req, res) => {
-    const { id } = req.params;
-    db.select('*').from('users').where({ id })
-        .then((user) => {
-            if (user.length) {
-                res.json(user[0])
-            }
-            else {
-                res.status(404).json('user not found');
-            }
-
-
-        }).catch((err) => {
-            res.status(500).json('Oops! Something is wrong');
-        });
-
-})
-
-app.put('/image', (req, res) => {
-
-    const { id } = req.body;
-    db('users').where('id', '=', id)
-        .increment('entries', 1)
-        .returning('entries')
-        .then(entries => {
-            res.json(entries[0]);
-        })
-        .catch(err => res.status(500).json('Unable to get entries'))
-
-
-})
+app.post('/signin', (req, res) => { signin.handleSignin(req, res, db, bcrypt) })
+app.post('/register', (req, res) => { register.handleRegister(req, res, db, bcrypt) })
+app.get('/profile/:id', (req, res) => { profile.handleProfile(req, res, db) })
+app.put('/image', (req, res) => { image.handleImage(req, res, db) })
+app.post('/imageurl', (req, res) => { image.handleApiCall(req, res) })
 
 
 app.listen(3005, () => {
